Multiclass_model-master/requirements.txt
PyYAML == 5.1.2
matplotlib == 3.1.1
numpy == 1.16.4
pandas == 0.24.2
pytest == 5.0.1
scikit_learn == 0.21.3
scipy == 1.3.1
shap == 0.30.1
xgboost == 0.90
-e
sphinx
dill
ruamel.yaml